Development

See merge request !213

+++ dbas/review/helper/history.py
@@ -9,14 +9,17 @@ from dbas.database import DBDiscussionSession
 from dbas.database.discussion_model import ReviewDelete, LastReviewerDelete, ReviewOptimization, \
     LastReviewerOptimization, User, ReputationHistory, ReputationReason, ReviewDeleteReason, ReviewEdit,\
     LastReviewerEdit, ReviewEditValue, TextVersion, Statement, ReviewCanceled, sql_timestamp_pretty_print,\
-    ReviewDuplicate, LastReviewerDuplicate, RevokedDuplicate, Argument, Premise
-from dbas.lib import get_text_for_argument_uid, get_profile_picture, is_user_author_or_admin, get_text_for_statement_uid
+    ReviewDuplicate, LastReviewerDuplicate, RevokedDuplicate, Argument, Premise, ReviewMerge, ReviewSplit,\
+    PremiseGroupMerged, PremiseGroupSplitted, LastReviewerSplit, LastReviewerMerge, ReviewSplitValues, \
+    ReviewMergeValues, StatementReplacementsByPremiseGroupSplit, StatementReplacementsByPremiseGroupMerge, \
+    ArgumentsAddedByPremiseGroupSplit
+from dbas.lib import get_text_for_argument_uid, get_profile_picture, is_user_author_or_admin, \
+    get_text_for_statement_uid, get_text_for_premisesgroup_uid
 from dbas.logger import logger
 from dbas.review.helper.main import en_or_disable_object_of_review
 from dbas.review.helper.reputation import get_reputation_of, reputation_borders, reputation_icons
 from dbas.strings.keywords import Keywords as _
 from dbas.strings.translator import Translator
-from sqlalchemy import and_
 
 
 def get_review_history(main_page, nickname, translator):
@@ -68,6 +71,8 @@ def __get_data(main_page, nickname, translator, is_executed=False):
     optimizations_list = __get_executed_reviews_of('optimizations', main_page, ReviewOptimization, LastReviewerOptimization, translator, is_executed)
     edits_list = __get_executed_reviews_of('edits', main_page, ReviewEdit, LastReviewerEdit, translator, is_executed)
     duplicates_list = __get_executed_reviews_of('duplicates', main_page, ReviewDuplicate, LastReviewerDuplicate, translator, is_executed)
+    splits_list = __get_executed_reviews_of('splits', main_page, ReviewSplit, LastReviewerSplit, translator, is_executed)
+    merges_list = __get_executed_reviews_of('merges', main_page, ReviewMerge, LastReviewerMerge, translator, is_executed)
 
     past_decision = [{
         'title': 'Delete Queue',
@@ -101,6 +106,22 @@ def __get_data(main_page, nickname, translator, is_executed=False):
         'has_reason': False,
         'has_oem_text': False,
         'has_duplicate_text': True
+    }, {
+        'title': 'Splits Queue',
+        'queue': 'splits',
+        'icon': reputation_icons['splits'],
+        'content': splits_list,
+        'has_reason': False,
+        'has_oem_text': True,
+        'has_duplicate_text': False
+    }, {
+        'title': 'Merges Queue',
+        'queue': 'merges',
+        'icon': reputation_icons['merges'],
+        'content': merges_list,
+        'has_reason': False,
+        'has_oem_text': True,
+        'has_duplicate_text': False
     }]
     ret_dict['past_decision'] = past_decision
 
@@ -186,6 +207,8 @@ def __get_executed_review_element_of(table, main_page, review, last_review_type,
     # getting text
     if table == 'duplicates':
         full_text = get_text_for_statement_uid(review.duplicate_statement_uid)
+    elif table in ['splits', 'merges']:
+        full_text, tmp = get_text_for_premisesgroup_uid(review.premisesgroup_uid)
     elif review.statement_uid is None:
         full_text = get_text_for_argument_uid(review.argument_uid)
     else:
@@ -201,11 +224,17 @@ def __get_executed_review_element_of(table, main_page, review, last_review_type,
     short_text += '...' if len(full_text) > length else '.'
     short_text = '<span class="text-primary">' + short_text + '</span>'
 
-    is_okay = False if table == 'optimizations' else True
-    # getting all pro and contra votes for this review
     all_votes = DBDiscussionSession.query(last_review_type).filter_by(review_uid=review.uid)
-    pro_votes = all_votes.filter_by(is_okay=is_okay).all()
-    con_votes = all_votes.filter(and_(last_review_type.is_okay != is_okay)).all()
+    is_okay = False if table == 'optimizations' else True
+    if table is 'merges':
+        pro_votes = all_votes.filter_by(should_merge=is_okay).all()
+        con_votes = all_votes.filter(last_review_type.should_merge != is_okay).all()
+    elif table is 'splits':
+        pro_votes = all_votes.filter_by(should_split=is_okay).all()
+        con_votes = all_votes.filter(last_review_type.should_split != is_okay).all()
+    else:
+        pro_votes = all_votes.filter_by(is_okay=is_okay).all()
+        con_votes = all_votes.filter(last_review_type.is_okay != is_okay).all()
 
     # getting the users which have voted
     pro_list = [__get_user_dict_for_review(pro.reviewer_uid, main_page) for pro in pro_votes]
@@ -283,6 +312,30 @@ def __handle_table_of_review_element(table, entry, review, short_text, full_text
         entry['statement_duplicate_shorttext'] = text[0:length] + ('...' if len(text) > length else '')
         entry['statement_duplicate_fulltext'] = text
 
+    if table is 'splits':
+        oem_fulltext, tmp = get_text_for_premisesgroup_uid(review.premisesgroup_uid)
+        full_text = oem_fulltext
+        db_values = DBDiscussionSession.query(ReviewSplitValues).filter_by(review_uid=review.uid).all()
+        if db_values:
+            full_text = str([value.content for value in db_values])
+        entry['argument_oem_shorttext'] = oem_fulltext[0:length] + '...' if len(oem_fulltext) > length else oem_fulltext
+        entry['argument_oem_fulltext'] = oem_fulltext
+        entry['argument_shorttext'] = full_text[0:length] + '...' if len(full_text) > length else full_text
+        entry['argument_fulltext'] = full_text
+
+    if table is 'merges':
+        db_premises = DBDiscussionSession.query(Premise).filter_by(premisesgroup_uid=review.premisesgroup_uid).all()
+        oem_fulltext = str([get_text_for_statement_uid(p.statement_uid) for p in db_premises])
+        full_text = oem_fulltext
+        db_values = DBDiscussionSession.query(ReviewMergeValues).filter_by(review_uid=review.uid).all()
+        if db_values:
+            full_text = str([value.content for value in db_values])
+        full_text = ' and '.join(full_text)
+        entry['argument_oem_shorttext'] = oem_fulltext[0:length] + '...' if len(oem_fulltext) > length else oem_fulltext
+        entry['argument_oem_fulltext'] = oem_fulltext
+        entry['argument_shorttext'] = full_text[0:length] + '...' if len(full_text) > length else full_text
+        entry['argument_fulltext'] = full_text
+
     return entry
 
 
@@ -331,17 +384,25 @@ def revoke_old_decision(queue, uid, lang, nickname):
     db_user = DBDiscussionSession.query(User).filter_by(nickname=nickname).first()
     _t = Translator(lang)
 
+    if not __is_uid_valid(uid, queue):
+        logger('review_history_helper', 'revoke_old_decision', 'no review with the uid or invalid queue: {},{}'.format(uid, queue), error=True)
+        error = _t.get(_.internalKeyError)
+        return success, error
+
     if queue == 'deletes':
+        logger('review_history_helper', 'revoke_old_decision', 'Executing deletes-queue')
         __revoke_decision_and_implications(ReviewDelete, LastReviewerDelete, uid)
         success = _t.get(_.dataRemoved)
         DBDiscussionSession.add(ReviewCanceled(author=db_user.uid, review_delete=uid))
 
     elif queue == 'optimizations':
+        logger('review_history_helper', 'revoke_old_decision', 'Executing optimizations-queue')
         __revoke_decision_and_implications(ReviewOptimization, LastReviewerOptimization, uid)
         success = _t.get(_.dataRemoved)
         DBDiscussionSession.add(ReviewCanceled(author=db_user.uid, review_optimization=uid))
 
     elif queue == 'edits':
+        logger('review_history_helper', 'revoke_old_decision', 'Executing edits-queue')
         db_review = DBDiscussionSession.query(ReviewEdit).get(uid)
         db_review.set_revoked(True)
         DBDiscussionSession.query(LastReviewerEdit).filter_by(review_uid=uid).delete()
@@ -360,6 +421,7 @@ def revoke_old_decision(queue, uid, lang, nickname):
         success = _t.get(_.dataRemoved)
 
     elif queue == 'duplicates':
+        logger('review_history_helper', 'revoke_old_decision', 'Executing duplicates-queue')
         db_review = DBDiscussionSession.query(ReviewDuplicate).get(uid)
         db_review.set_revoked(True)
         DBDiscussionSession.add(ReviewCanceled(author=db_user.uid, review_duplicate=uid))
@@ -367,13 +429,93 @@ def revoke_old_decision(queue, uid, lang, nickname):
 
         success = _t.get(_.dataRemoved)
 
+    elif queue == 'merges':
+        logger('review_history_helper', 'revoke_old_decision', 'Executing merges-queue')
+        db_review = DBDiscussionSession.query(ReviewMerge).get(uid)
+        db_review.set_revoked(True)
+        db_pgroup_merged = DBDiscussionSession.query(PremiseGroupMerged).filter_by(review_uid=uid).all()
+        replacements = DBDiscussionSession.query(StatementReplacementsByPremiseGroupMerge).filter_by(review_uid=uid).all()
+        __undo_premisegroups(db_pgroup_merged, replacements)
+
+        DBDiscussionSession.query(LastReviewerSplit).filter_by(review_uid=uid).delete()
+        DBDiscussionSession.query(ReviewSplitValues).filter_by(review_uid=uid).delete()
+        DBDiscussionSession.query(StatementReplacementsByPremiseGroupMerge).filter_by(review_uid=uid).delete()
+
+        success = _t.get(_.dataRemoved)
+
+    elif queue == 'splits':
+        logger('review_history_helper', 'revoke_old_decision', 'Executing splits-queue')
+        db_review = DBDiscussionSession.query(ReviewSplit).get(uid)
+        db_review.set_revoked(True)
+        db_pgroup_splitted = DBDiscussionSession.query(PremiseGroupSplitted).filter_by(review_uid=uid).all()
+        replacements = DBDiscussionSession.query(StatementReplacementsByPremiseGroupSplit).filter_by(review_uid=uid).all()
+        disable_args = [arg.uid for arg in DBDiscussionSession.query(ArgumentsAddedByPremiseGroupSplit).filter_by(review_uid=uid).all()]
+        __undo_premisegroups(db_pgroup_splitted, replacements)
+        __disable_arguments_by_id(disable_args)
+
+        DBDiscussionSession.query(LastReviewerMerge).filter_by(review_uid=uid).delete()
+        DBDiscussionSession.query(ReviewMergeValues).filter_by(review_uid=uid).delete()
+        DBDiscussionSession.query(StatementReplacementsByPremiseGroupSplit).filter_by(review_uid=uid).delete()
+
+        success = _t.get(_.dataRemoved)
+
     else:
+        logger('review_history_helper', 'revoke_old_decision', 'no queue found: {},{}'.format(uid, queue), error=True)
         error = _t.get(_.internalKeyError)
 
     DBDiscussionSession.flush()
     transaction.commit()
 
     return success, error
+
+
+def __disable_arguments_by_id(argument_uids):
+    """
+    Disbale the list of argument by their id
+
+    :param Argument_uids: list of argument.uid
+    :return: None
+    """
+    for uid in argument_uids:
+        db_arg = DBDiscussionSession.query(Argument).get(uid)
+        db_arg.set_disable(True)
+        DBDiscussionSession.add(db_arg)
+        DBDiscussionSession.flush()
+
+
+def __undo_premisegroups(pgroups_splitted_or_merged, replacements):
+    """
+
+    :param pgroups_splitted_or_merged:
+    :param replacements:
+    :return:
+    """
+    logger('review_history_helper', '__undo_premisegroups', 'Got {} merge/splitted pgroups and {} replacements'.format(len(pgroups_splitted_or_merged), len(replacements)))
+
+    for element in pgroups_splitted_or_merged:
+        old_pgroup = element.old_premisegroup_uid
+        new_pgroup = element.new_premisegroup_uid
+
+        db_arguments = DBDiscussionSession.query(Argument).filter_by(premisesgroup_uid=new_pgroup).all()
+        for argument in db_arguments:
+            logger('review_history_helper', '__undo_premisegroups', 'reset arguments {} pgroup from {} back to {}'.format(argument.uid, new_pgroup, old_pgroup))
+            argument.set_premisegroup(old_pgroup)
+            DBDiscussionSession.add(argument)
+            DBDiscussionSession.flush()
+
+    for element in replacements:
+        old_statement = element.old_statement_uid
+        new_statement = element.new_statement_uid
+
+        db_arguments = DBDiscussionSession.query(Argument).filter_by(conclusion_uid=new_statement).all()
+        for argument in db_arguments:
+            logger('review_history_helper', '__undo_premisegroups', 'reset arguments {} conclusion from {} back to {}'.format(argument.uid, new_statement, old_statement))
+            argument.set_conclusion(old_statement)
+            DBDiscussionSession.add(argument)
+            DBDiscussionSession.flush()
+
+    DBDiscussionSession.flush()
+    transaction.commit()
 
 
 def cancel_ongoing_decision(queue, uid, lang, nickname):
@@ -390,8 +532,11 @@ def cancel_ongoing_decision(queue, uid, lang, nickname):
     success = ''
     error = ''
     db_user = DBDiscussionSession.query(User).filter_by(nickname=nickname).first()
-
     _t = Translator(lang)
+
+    if not __is_uid_valid(uid, queue):
+        return success, _t.get(_.internalKeyError)
+
     if queue == 'deletes':
         DBDiscussionSession.query(ReviewDelete).get(uid).set_revoked(True)
         DBDiscussionSession.query(LastReviewerDelete).filter_by(review_uid=uid).delete()
@@ -417,6 +562,18 @@ def cancel_ongoing_decision(queue, uid, lang, nickname):
         success = _t.get(_.dataRemoved)
         DBDiscussionSession.add(ReviewCanceled(author=db_user.uid, review_duplicate=uid, was_ongoing=True))
 
+    elif queue == 'merges':
+        DBDiscussionSession.query(ReviewMerge).get(uid).set_revoked(True)
+        DBDiscussionSession.query(LastReviewerMerge).filter_by(review_uid=uid).delete()
+        DBDiscussionSession.query(PremiseGroupMerged).filter_by(review_uid=uid).delete()
+        success = _t.get(_.dataRemoved)
+
+    elif queue == 'splits':
+        DBDiscussionSession.query(ReviewSplit).get(uid).set_revoked(True)
+        DBDiscussionSession.query(LastReviewerSplit).filter_by(review_uid=uid).delete()
+        DBDiscussionSession.query(PremiseGroupSplitted).filter_by(review_uid=uid).delete()
+        success = _t.get(_.dataRemoved)
+
     else:
         error = _t.get(_.internalKeyError)
 
@@ -424,6 +581,33 @@ def cancel_ongoing_decision(queue, uid, lang, nickname):
     transaction.commit()
 
     return success, error
+
+
+def __is_uid_valid(uid, queue):
+    """
+    Check for the specific review in the fiven queue
+
+    :param queue: Table of review
+    :param uid: Review.uid
+    :return: Boolean
+    :rtype: Boolean
+    """
+
+    mapping = {
+        'deletes': ReviewDelete,
+        'optimizations': ReviewOptimization,
+        'edits': ReviewEdit,
+        'duplicates': ReviewDuplicate,
+        'merges': ReviewMerge,
+        'splits': ReviewSplit,
+    }
+
+    if queue in mapping:
+        # logger('review_history_helper', '__is_uid_valid', 'query table {} with uid {}'.format(mapping[queue], uid))
+        return DBDiscussionSession.query(mapping[queue]).get(uid) is not None
+
+    # logger('review_history_helper', '__is_uid_valid', 'no table found for {}'.format(queue), error=True)
+    return False
 
 
 def __revoke_decision_and_implications(type, reviewer_type, uid):
